Fix script tx to be spent on L1
- Add missing collateral
- Add missing witness during signing

Note:
By introducing the collateral utxo, now the key has more than one.
This causes the resulting utxo from queryUTxOFor contained two entries: one for
the normal utxo and another one for the collateral utxo.
This was making createScriptOutput to build an invalid script tx in case the
collateral utxo was selected, when finding a utxo that fits the total amount
to deposit.

Finally, removed debug lines